requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
pydantic==1.10.13 
langchain>=0.1.17
langchain-google-genai>=0.0.7
google-generativeai>=0.3.2